ai_agents_platform/evaluation_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langsmith
openai
psycopg2-binary
sqlalchemy
python-multipart
pydub
openai-whisper
